small improvements according to the PR comments
