+++ src/sparseml/pytorch/utils/helpers.py
@@ -85,12 +85,12 @@ __all__ = [
     "tensor_sample",
     "mask_difference",
     "get_layer",
-    "replace_layer",
     "get_terminal_layers",
     "get_conv_layers",
     "get_linear_layers",
     "get_prunable_layers",
     "get_quantizable_layers",
+    "swap_modules",
     "get_named_layers_and_params_by_regex",
     "any_str_or_regex_matches_param_name",
     "NamedLayerParam",
@@ -724,32 +724,6 @@ def get_layer(name: str, module: Module) -> Module:
 
     return layer
 
-
-def replace_layer(
-    module: Module,
-    name: str,
-    replace: Module,
-) -> Module:
-    """
-    General function to replace a layer in a module with the given new one.
-
-    :param module: the module to replace the layer in
-    :param name: the name of the layer to replace the activation for
-    :param replace: the module to replace the layer with
-    :return: the original layer that was replaced
-    """
-    parent = module
-    sections = name.split(".")
-
-    for sec in sections[:-1]:
-        parent = parent.__getattr__(sec)
-
-    cur = parent.__getattr__(sections[-1])
-    parent.__setattr__(sections[-1], replace)
-
-    return cur
-
-
 def get_terminal_layers(module: Module) -> Dict[str, Module]:
     """
     :param module: the module to grab all terminal layers for
@@ -1248,3 +1222,50 @@ def get_dependency_order(
     for h in handles:
         h.remove()
     return order
+
+
+def swap_modules(
+    module: torch.nn.Module, submodule_name: str, submodule_to_replace: torch.nn.Module
+):
+    """
+    Recursively unfold the submodules of the module according to the submodule_name
+    to eventually replace the leaf submodule (accessed from the module through the
+    submodule_name) with the submodule_to_replace.
+
+    E.g
+    ```
+    swap_modules(module=Model,
+                 module_name="layers.0.sublayer",
+                 module_to_replace=ReplaceModule
+                 )
+    ```
+    this will recursively call:
+    1. SomeModule1 = getattr(Model, "layers")
+    2. SomeModule2 = getattr(SomeModule1, "0")
+
+    finally will swap the leaf submodule with the submodule_to_replace
+    ```
+    (submodule_name = "sublayer")
+    setattr(SomeModule2 , submodule_name, ReplaceModule)
+    ```
+    this will essentially replace SomeModule2.sublayer with ReplaceModule
+
+    :param module: the module to replace with the module_to_replace
+    :param submodule_name: the name of the module to replace
+    :param submodule_to_replace: the module to replace the module with
+    """
+    if not isinstance(module, torch.nn.Module):
+        raise ValueError(f"module {module} is not a torch.nn.Module")
+    if not isinstance(submodule_to_replace, torch.nn.Module):
+        raise ValueError(
+            f"submodule_to_replace {submodule_name} is not a torch.nn.Module"
+        )
+
+    attribute_name = submodule_name
+    attribute_name = attribute_name.split(".", 1)
+    if len(attribute_name) == 1:
+        setattr(module, attribute_name[0], submodule_to_replace)
+    else:
+        swap_modules(
+            getattr(module, attribute_name[0]), attribute_name[1], submodule_to_replace
+        )
