+++ src/sparseml/transformers/sparsification/modification/modification_objects.py
@@ -23,58 +23,10 @@ import torch
 __all__ = [
     "QuantizableIdentity",
     "QuantizableMatMul",
-    "swap_modules",
     "QuantizableBatchMatmul",
     "QATMatMul",
     "QATLinear",
 ]
-
-
-def swap_modules(
-    module: torch.nn.Module, submodule_name: str, submodule_to_replace: torch.nn.Module
-):
-    """
-    Recursively unfold the submodules of the module according to the submodule_name
-    to eventually replace the leaf submodule (accessed from the module through the
-    submodule_name) with the submodule_to_replace.
-
-    E.g
-    ```
-    swap_modules(module=Model,
-                 module_name="layers.0.sublayer",
-                 module_to_replace=ReplaceModule
-                 )
-    ```
-    this will recursively call:
-    1. SomeModule1 = getattr(Model, "layers")
-    2. SomeModule2 = getattr(SomeModule1, "0")
-
-    finally will swap the leaf submodule with the submodule_to_replace
-    ```
-    (submodule_name = "sublayer")
-    setattr(SomeModule2 , submodule_name, ReplaceModule)
-    ```
-    this will essentially replace SomeModule2.sublayer with ReplaceModule
-
-    :param module: the module to replace with the module_to_replace
-    :param submodule_name: the name of the module to replace
-    :param submodule_to_replace: the module to replace the module with
-    """
-    if not isinstance(module, torch.nn.Module):
-        raise ValueError(f"module {module} is not a torch.nn.Module")
-    if not isinstance(submodule_to_replace, torch.nn.Module):
-        raise ValueError(
-            f"submodule_to_replace {submodule_name} is not a torch.nn.Module"
-        )
-
-    attribute_name = submodule_name
-    attribute_name = attribute_name.split(".", 1)
-    if len(attribute_name) == 1:
-        setattr(module, attribute_name[0], submodule_to_replace)
-    else:
-        swap_modules(
-            getattr(module, attribute_name[0]), attribute_name[1], submodule_to_replace
-        )
 
 
 class QuantizableIdentity(torch.nn.Module):
