+++ src/sparseml/transformers/sparsification/modification/modifying_mistral.py
@@ -19,11 +19,13 @@ context of SparseML
 import logging
 import math
 from typing import Optional, Tuple
+from sparseml.pytorch.utils.helpers import swap_modules
 
 import torch
 from torch import nn
 from transformers.models.mistral.modeling_mistral import (
     MistralAttention,
+    MistralFlashAttention2,
     apply_rotary_pos_emb,
     repeat_kv,
 )
@@ -31,7 +33,6 @@ from transformers.models.mistral.modeling_mistral import (
 from sparseml.transformers.sparsification.modification.modification_objects import (
     QuantizableIdentity,
     QuantizableMatMul,
-    swap_modules,
 )
 from sparseml.transformers.sparsification.modification.registry import (
     ModificationRegistry,
@@ -56,12 +57,11 @@ def modify(model: torch.nn.Module) -> torch.nn.Module:
     :return: the modified Mistral model
     """
     for name, submodule in model.named_modules():
-        submodule_cname = submodule.__class__.__name__
-        if submodule_cname == "MistralAttention":
+        if isinstance(submodule, MistralAttention):
             swap_modules(model, name, MistralAttentionWithQuantizableMatmuls(submodule))
-        if submodule_cname == "MistralFlashAttention2":
+        if isinstance(submodule, MistralFlashAttention2):
             _LOGGER.debug(
-                f"The model contains {submodule_cname} "
+                f"The model contains {submodule.__class__.__name__} "
                 "module, which will not be modified"
             )
     return model
